requirements.txt

groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
